requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.7.0
# Requerimientos caja
xhtml2pdf==0.2.17
weasyprint==65.0

# Requerimientos para la Conexion de Android y Django
djangorestframework-simplejwt==5.5.0
